move write_vis into contrib
